+++ superset/app.py
@@ -146,11 +146,11 @@ class SupersetAppInitializer:
         )
         from superset.views.api import Api
         from superset.views.core import (
-            AccessRequestsModelView,
             KV,
             R,
             Superset,
         )
+        from superset.views.access_requests import AccessRequestsModelView
         from superset.views.css_templates import CssTemplateAsyncModelView
         from superset.views.css_templates import CssTemplateModelView
         from superset.charts.api import ChartRestApi
